+++ AdventOfCode2021/AdventOfCode2021.sln
@@ -13,7 +13,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DayTwo", "DayTwo\DayTwo.csp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DayThree", "DayThree\DayThree.csproj", "{627FAC57-D00F-4835-8D4B-B4AFD2AB14EB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DayThree.Test", "DayThree.Test\DayThree.Test.csproj", "{D6EA339A-D605-4651-915B-0A1CB6797B28}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Solutions.Test", "DayThree.Test\Solutions.Test.csproj", "{D6EA339A-D605-4651-915B-0A1CB6797B28}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
