+++ AdventOfCode2021/DayThree.Test/DayThree/PartTwoTests.cs
@@ -1,7 +1,8 @@
-﻿using FluentAssertions;
+﻿using DayThree;
+using FluentAssertions;
 using Xunit;
 
-namespace DayThree.Test;
+namespace Solutions.Test.DayThree;
 
 public class PartTwoTests
 {
