+++ AdventOfCode2021/DayThree.Test/Solutions.Test.csproj
